+++ src/common/utils/deployUtils.ts
@@ -461,23 +461,25 @@ async function buildDeployOncePackageXml(debugMode = false, options: any = {}) {
     packageNoOverwrite = path.resolve("./manifest/packageDeployOnce.xml");
   }
   if (fs.existsSync(packageNoOverwrite)) {
-    uxLog(this, "Building package-no-overwrite.xml...");
+    uxLog(this, c.cyan("Handling package-no-overwrite.xml..."));
     // If package-no-overwrite.xml is not empty, build target org package.xml and remove its content from packageOnce.xml
     if (!(await isPackageXmlEmpty(packageNoOverwrite))) {
       const tmpDir = await createTempDir();
       // Build target org package.xml
-      uxLog(this, c.cyan(`Generating full package.xml from target org to remove its content matching package-no-overwrite.xml ...`));
+      uxLog(this, c.cyan(`Generating full package.xml from target org to identify its items matching with package-no-overwrite.xml ...`));
       const targetOrgPackageXml = path.join(tmpDir, "packageTargetOrg.xml");
       await buildOrgManifest(options.targetUsername, targetOrgPackageXml, options.conn);
 
-      const packageNoOverwriteToUse = path.join(tmpDir, "package-no-overwrite.xml");
-      await fs.copy(packageNoOverwrite, packageNoOverwriteToUse);
+      let calculatedPackageNoOverwrite = path.join(tmpDir, "package-no-overwrite.xml");
+      await fs.copy(packageNoOverwrite, calculatedPackageNoOverwrite);
       // Keep in deployOnce.xml only what is necessary to deploy
-      await removePackageXmlContent(packageNoOverwriteToUse, targetOrgPackageXml, true, { debugMode: debugMode, keepEmptyTypes: false });
-      uxLog(this, c.grey(`package-no-overwrite.xml with only metadatas that do not exist in target: ${packageNoOverwriteToUse}`));
-      // Check if there is still something in updated package-no-overwrite.xml
-      if (!(await isPackageXmlEmpty(packageNoOverwriteToUse))) {
-        return packageNoOverwriteToUse;
+      await removePackageXmlContent(calculatedPackageNoOverwrite, targetOrgPackageXml, true, { debugMode: debugMode, keepEmptyTypes: false });
+      await fs.copy(calculatedPackageNoOverwrite, path.join(tmpDir, "calculated-package-no-overwrite.xml"));
+      calculatedPackageNoOverwrite = path.join(tmpDir, "calculated-package-no-overwrite.xml");
+      uxLog(this, c.grey(`calculated-package-no-overwrite.xml with only items that already exist in target org: ${calculatedPackageNoOverwrite}`));
+      // Check if there is still something in calculated-package-no-overwrite.xml
+      if (!(await isPackageXmlEmpty(calculatedPackageNoOverwrite))) {
+        return calculatedPackageNoOverwrite;
       }
     }
   }
@@ -550,12 +552,12 @@ export async function removePackageXmlContent(
   options = { debugMode: false, keepEmptyTypes: false },
 ) {
   if (removedOnly === false) {
-    uxLog(this, c.cyan(`Removing ${c.green(path.basename(packageXmlFileToRemove))} content from ${c.green(path.basename(packageXmlFile))}...`));
+    uxLog(this, c.cyan(`Removing ${c.green(path.basename(packageXmlFileToRemove))} items from ${c.green(path.basename(packageXmlFile))}...`));
   } else {
     uxLog(
       this,
       c.cyan(
-        `Keeping ${c.green(path.basename(packageXmlFileToRemove))} content from ${c.green(path.basename(packageXmlFile))} (and remove the rest)...`,
+        `Keeping ${c.green(path.basename(packageXmlFileToRemove))} items matching with ${c.green(path.basename(packageXmlFile))} (and remove the rest)...`,
       ),
     );
   }