+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 ## [4.23.2] 2023-12-25
 
 - Catch errors when a JIRA comment has not been posted
-- Reduce package-no-overwrite.xml management logs
+- Simplify package-no-overwrite.xml management logs
 
 ## [4.23.1] 2023-12-25